renamet mocks.ts til use-analytics-test-mocks.ts
